[FIX] adjusment request increase decrease

+++ app/Http/Controllers/Api/AdjustmentRequestController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\AdjustmentRequestStoreRequest;
 use App\Http\Resources\AdjustmentRequestResource;
 use App\Models\AdjustmentRequest;
+use App\Models\Stock;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
@@ -78,8 +79,11 @@ class AdjustmentRequestController extends Controller
     {
         // abort_if(!auth('sanctum')->user()->tokenCan('adjustment_request_approve'), 403);
 
+        if (!is_null($adjustmentRequest->is_approved)) return response()->json(['message' => sprintf("Adjustment request sudah di %s. Tidak dapat di edit!", $adjustmentRequest->is_approved ? 'approve' : 'reject')], 404);
+
+        /** @var \App\Models\StockProductUnit $stockProductUnit */
         $stockProductUnit = $adjustmentRequest->stockProductUnit;
-        if (!$stockProductUnit) return response()->json(['message' => "Stock product unit Tidak ditemukan"], 404);
+        if (!$stockProductUnit) return response()->json(['message' => "Stock product unit tidak ditemukan"], 404);
 
         DB::beginTransaction();
         try {
@@ -88,26 +92,33 @@ class AdjustmentRequestController extends Controller
             $adjustmentRequest->approved_by = auth('sanctum')->id();
             $adjustmentRequest->approved_datetime = now();
 
-            if ($adjustmentRequest->isDirty('is_approved')) {
-                if ($adjustmentRequest->is_approved == 1) {
+            if ($adjustmentRequest->is_approved == true) {
+                if ($adjustmentRequest->is_increment == 1) {
                     $folder = 'qrcode/';
-                    for ($i = 0; $i < $adjustmentRequest->value ?? 0; $i++) {
-                        $stock = $stockProductUnit->stocks()->create([
-                            'adjustment_request_id' => $adjustmentRequest->id
-                        ]);
 
-                        // $logo = public_path('images/logo-platinum.png');
+                    if ($stockProductUnit->productUnit->is_generate_qr) {
+                        // GenerateStockQrcode::dispatch($stockProductUnit, $qty, $folder);
 
-                        $data = QrCode::size(350)
-                            ->format('png')
-                            // ->merge($logo, absolute: true)
-                            ->generate($stock->id);
+                        for ($i = 0; $i < $adjustmentRequest->value ?? 0; $i++) {
+                            $stock = $stockProductUnit->stocks()->create([
+                                'adjustment_request_id' => $adjustmentRequest->id
+                            ]);
 
-                        $fileName = $adjustmentRequest->id . '/' . $stock->id . '.png';
-                        $fullPath = $folder .  $fileName;
-                        Storage::put($fullPath, $data);
+                            // $logo = public_path('images/logo-platinum.png');
 
-                        $stock->update(['qr_code' => $fullPath]);
+                            $data = QrCode::size(350)
+                                ->format('png')
+                                // ->merge($logo, absolute: true)
+                                ->generate($stock->id);
+
+                            $fileName = $adjustmentRequest->id . '/' . $stock->id . '.png';
+                            $fullPath = $folder .  $fileName;
+                            Storage::put($fullPath, $data);
+
+                            $stock->update(['qr_code' => $fullPath]);
+                        }
+                    } else {
+                        $stockProductUnit->increment('qty', $adjustmentRequest->value);
                     }
 
                     $adjustmentRequest->histories()->create([
@@ -115,26 +126,31 @@ class AdjustmentRequestController extends Controller
                         'stock_product_unit_id' => $adjustmentRequest->stock_product_unit_id,
                         'value' => $adjustmentRequest->value ?? 0,
                         'is_increment' => 1,
-                        'description' => 'Adjustment request - ' . $adjustmentRequest->description,
+                        'description' => 'Adjustment request (Penambahan) - ' . $adjustmentRequest->description,
                         'ip' => request()->ip(),
                         'agent' => request()->header('user-agent'),
                     ]);
                 } else {
-                    $adjustmentRequest->stocks?->each->forceDelete();
+                    if ($stockProductUnit->productUnit->is_generate_qr) {
+                        $adjustmentRequest->stocks?->each->forceDelete();
+                        Stock::whereAvailableStock()->where('stock_product_unit_id', $stockProductUnit->id)->limit($adjustmentRequest->value)->delete();
 
-                    if ($adjustmentRequest->getOriginal('is_approved')) {
-                        $adjustmentRequest->histories()->create([
-                            'user_id' => auth('sanctum')->id(),
-                            'stock_product_unit_id' => $adjustmentRequest->stock_product_unit_id,
-                            'value' => $adjustmentRequest->value ?? 0,
-                            'is_increment' => 0,
-                            'description' => 'Adjustment request - ' . $adjustmentRequest->description,
-                            'ip' => request()->ip(),
-                            'agent' => request()->header('user-agent'),
-                        ]);
+                        // Storage::deleteDirectory($adjustmentRequest->id);
+                    } else {
+                        $stockProductUnit->decrement('qty', $adjustmentRequest->value);
                     }
 
-                    Storage::deleteDirectory($adjustmentRequest->id);
+                    // if ($adjustmentRequest->getOriginal('is_approved')) {
+                    $adjustmentRequest->histories()->create([
+                        'user_id' => auth('sanctum')->id(),
+                        'stock_product_unit_id' => $adjustmentRequest->stock_product_unit_id,
+                        'value' => $adjustmentRequest->value ?? 0,
+                        'is_increment' => 0,
+                        'description' => 'Adjustment request (Pengurangan) - ' . $adjustmentRequest->description,
+                        'ip' => request()->ip(),
+                        'agent' => request()->header('user-agent'),
+                    ]);
+                    // }
                 }
             }
 
@@ -148,4 +164,79 @@ class AdjustmentRequestController extends Controller
         $message = 'Data ' . ($adjustmentRequest->is_approved ? 'approved' : 'rejected') . ' successfully';
         return response()->json(['message' => $message]);
     }
+
+    // public function approve(AdjustmentRequest $adjustmentRequest, Request $request)
+    // {
+    //     // abort_if(!auth('sanctum')->user()->tokenCan('adjustment_request_approve'), 403);
+
+    //     $stockProductUnit = $adjustmentRequest->stockProductUnit;
+    //     if (!$stockProductUnit) return response()->json(['message' => "Stock product unit Tidak ditemukan"], 404);
+
+    //     DB::beginTransaction();
+    //     try {
+    //         $adjustmentRequest->reason = $request->reason ?? null;
+    //         $adjustmentRequest->is_approved = $request->is_approved ?? null;
+    //         $adjustmentRequest->approved_by = auth('sanctum')->id();
+    //         $adjustmentRequest->approved_datetime = now();
+
+    //         if ($adjustmentRequest->isDirty('is_approved')) {
+    //             if ($adjustmentRequest->is_approved == 1) {
+    //                 $folder = 'qrcode/';
+    //                 for ($i = 0; $i < $adjustmentRequest->value ?? 0; $i++) {
+    //                     $stock = $stockProductUnit->stocks()->create([
+    //                         'adjustment_request_id' => $adjustmentRequest->id
+    //                     ]);
+
+    //                     // $logo = public_path('images/logo-platinum.png');
+
+    //                     $data = QrCode::size(350)
+    //                         ->format('png')
+    //                         // ->merge($logo, absolute: true)
+    //                         ->generate($stock->id);
+
+    //                     $fileName = $adjustmentRequest->id . '/' . $stock->id . '.png';
+    //                     $fullPath = $folder .  $fileName;
+    //                     Storage::put($fullPath, $data);
+
+    //                     $stock->update(['qr_code' => $fullPath]);
+    //                 }
+
+    //                 $adjustmentRequest->histories()->create([
+    //                     'user_id' => auth('sanctum')->id(),
+    //                     'stock_product_unit_id' => $adjustmentRequest->stock_product_unit_id,
+    //                     'value' => $adjustmentRequest->value ?? 0,
+    //                     'is_increment' => 1,
+    //                     'description' => 'Adjustment request - ' . $adjustmentRequest->description,
+    //                     'ip' => request()->ip(),
+    //                     'agent' => request()->header('user-agent'),
+    //                 ]);
+    //             } else {
+    //                 $adjustmentRequest->stocks?->each->forceDelete();
+
+    //                 if ($adjustmentRequest->getOriginal('is_approved')) {
+    //                     $adjustmentRequest->histories()->create([
+    //                         'user_id' => auth('sanctum')->id(),
+    //                         'stock_product_unit_id' => $adjustmentRequest->stock_product_unit_id,
+    //                         'value' => $adjustmentRequest->value ?? 0,
+    //                         'is_increment' => 0,
+    //                         'description' => 'Adjustment request - ' . $adjustmentRequest->description,
+    //                         'ip' => request()->ip(),
+    //                         'agent' => request()->header('user-agent'),
+    //                     ]);
+    //                 }
+
+    //                 Storage::deleteDirectory($adjustmentRequest->id);
+    //             }
+    //         }
+
+    //         $adjustmentRequest->save();
+    //         DB::commit();
+    //     } catch (\Throwable $th) {
+    //         DB::rollBack();
+    //         return response()->json(['message' => $th->getMessage()], 500);
+    //     }
+
+    //     $message = 'Data ' . ($adjustmentRequest->is_approved ? 'approved' : 'rejected') . ' successfully';
+    //     return response()->json(['message' => $message]);
+    // }
 }
